move reusable codes back to IpcServiceEndpoint from NamedPipeIpcServiceEndpoint
